moved Converter<Set<Enum<*>>> to Converters, not using SharedSecrets which is inaccessible in JDK 9
